version 1 test

+++ src/main/java/com/example/demo/repository/EmployeeRepository.java
@@ -3,6 +3,7 @@ package com.example.demo.repository;
 import com.example.demo.domain.Employee;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EmployeeRepository extends CrudRepository<Employee, Long> {
+public interface EmployeeRepository extends CrudRepository<Employee, Integer> {
+
 
 }
